+++ tests/test_x_cli.py
@@ -161,3 +161,269 @@ class TestCli:
             self.cli.run_command()
             captured = capsys.readouterr()
             assert " Online (online): True" in captured.out
+
+    def test_api_deviceBaseList(self, capsys):  # pylint: disable=invalid-name
+        """Test api_deviceBaseList: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "deviceBaseList", "deviceBaseList_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_deviceBaseList",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "8L0DF93PAZ55FD2" in captured.out
+
+    def test_api_deviceOpenList(self, capsys):  # pylint: disable=invalid-name
+        """Test api_deviceOpenList: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "deviceOpenList", "deviceOpenList_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_deviceOpenList",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "8L0DF93PAZ55FD2" in captured.out
+
+    def test_api_deviceBaseDetailList(self, capsys):  # pylint: disable=invalid-name
+        """Test api_deviceBaseDetailList: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "deviceBaseDetailList", "deviceBaseDetailList_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_deviceBaseDetailList",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "8L0DF93PAZ55FD2" in captured.out
+
+    def test_api_deviceOpenDetailList(self, capsys):  # pylint: disable=invalid-name
+        """Test api_deviceOpenDetailList: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "deviceOpenDetailList", "deviceOpenDetailList_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_deviceOpenDetailList",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "8L0DF93PAZ55FD2" in captured.out
+
+    def test_api_listDeviceAbility(self, capsys):  # pylint: disable=invalid-name
+        """Test api_listDeviceAbility: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "listDeviceAbility", "listDeviceAbility_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_listDeviceAbility",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "WLAN" in captured.out
+
+    def test_api_getAlarmMessage(self, capsys):  # pylint: disable=invalid-name
+        """Test api_getAlarmMessage: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "getAlarmMessage", "getAlarmMessage_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_getAlarmMessage",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "human" in captured.out
+
+    def test_api_deviceStorage(self, capsys):  # pylint: disable=invalid-name
+        """Test api_deviceStorage: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "deviceStorage", "deviceStorage_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_deviceStorage",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "totalBytes" in captured.out
+
+    def test_api_getDeviceCameraStatus(self, capsys):  # pylint: disable=invalid-name
+        """Test api_getDeviceCameraStatus: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "getDeviceCameraStatus", "getDeviceCameraStatus_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_getDeviceCameraStatus",
+                "device_id",
+                "headerDetect",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "headerDetect" in captured.out
+
+    def test_api_setDeviceCameraStatus(self, capsys):  # pylint: disable=invalid-name
+        """Test api_setDeviceCameraStatus: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "setDeviceCameraStatus", "setDeviceCameraStatus_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_setDeviceCameraStatus",
+                "device_id",
+                "headerDetect",
+                "on",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "{}" in captured.out
+
+    def test_api_getNightVisionMode(self, capsys):  # pylint: disable=invalid-name
+        """Test api_getNightVisionMode: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "getNightVisionMode", "getNightVisionMode_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_getNightVisionMode",
+                "device_id",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "Intelligent" in captured.out
+
+    def test_api_setNightVisionMode(self, capsys):  # pylint: disable=invalid-name
+        """Test api_setNightVisionMode: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "setNightVisionMode", "setNightVisionMode_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_setNightVisionMode",
+                "device_id",
+                "Intelligent",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "{}" in captured.out
+
+    def test_api_getMessageCallback(self, capsys):  # pylint: disable=invalid-name
+        """Test api_getMessageCallback: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "getMessageCallback", "getMessageCallback_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_getMessageCallback",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "callbackUrl" in captured.out
+
+    def test_api_setMessageCallbackOn(self, capsys):  # pylint: disable=invalid-name
+        """Test api_setMessageCallbackOn: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "setMessageCallback", "setMessageCallbackOn_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_setMessageCallbackOn",
+                "https://url.com",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "{}" in captured.out
+
+    def test_api_setMessageCallbackOff(self, capsys):  # pylint: disable=invalid-name
+        """Test api_setMessageCallbackOff: ok."""
+        with aioresponses() as mocked:
+            self.config_mock(mocked, "accessToken", "accessToken_ok")
+            self.config_mock(mocked, "setMessageCallback", "setMessageCallbackOff_ok")
+            self.cli.argv = [
+                "cli",
+                "--app-id",
+                "app_id",
+                "--app-secret",
+                "app_secret",
+                "api_setMessageCallbackOff",
+            ]
+            self.cli.parse_command_line()
+            self.cli.run_command()
+            captured = capsys.readouterr()
+            assert "{}" in captured.out